Will only work on campus. dbUserProfile.php needs to be modified. User -> UserProfile
ConfirmCancel - The SW had canceled the request. An email sent to the SW from the RMH staff saying that the request for canceling a reservation has been received and updated.
ModifyAccept and ModifyDeny - the modification request for a reservation made by the SW has been modified and then accepted or rejected.