Issue 74
Changed first field to patient last name (makes more logical sense)
Fixed spacing, put results in a table. 
Added "Parent" or "Patient" to the header so it's more clear what it's referring to.
Added a patient diagnosis section so the social worker can confirm symptoms/notes.
Added error message if the security check failed.
Added error message if family searched is not found with an option to create a family.

 -Jesse